requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
transformers>=4.30.0
torch>=2.0.0
Pillow>=10.0.0

